Split up testing/integration/internal
This moves testing/integration/internal into several more specific
packages to prepare to refactor the rest and dedup with the main envtest
code.
