+++ pkg/internal/testing/process/bin_path_finder_test.go
@@ -1,4 +1,4 @@
-package internal
+package process
 
 import (
 	"os"
