+++ pkg/internal/testing/process/internal_suite_test.go
@@ -1,0 +1,17 @@
+package process_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
+)
+
+func TestInternal(t *testing.T) {
+	t.Parallel()
+	RegisterFailHandler(Fail)
+	suiteName := "Envtest Process Launcher Suite"
+	RunSpecsWithDefaultAndCustomReporters(t, suiteName, []Reporter{printer.NewlineReporter{}, printer.NewProwReporter(suiteName)})
+}
